chore: Share Application initialization
