--- src/Uno.UI/UI/Xaml/Application.reference.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using Windows.Globalization;
-
-namespace Windows.UI.Xaml;
-
-public partial class Application
-{
-	public Application()
-	{
-		Current = this;
-		ApplicationLanguages.ApplyCulture();
-		InitializeSystemTheme();
-	}
-}
